# openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/customers/": {
        "get": {
          "tags": [
            "Customers"
          ],
          "summary": "Get Customers",
          "operationId": "get_customers_customers__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/CustomerRead"
                    },
                    "type": "array",
                    "title": "Response Get Customers Customers  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Customers"
          ],
          "summary": "Create Customer",
          "operationId": "create_customer_customers__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/customers/{customer_id}": {
        "get": {
          "tags": [
            "Customers"
          ],
          "summary": "Get Customer",
          "operationId": "get_customer_customers__customer_id__get",
          "parameters": [
            {
              "name": "customer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Customer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Customers"
          ],
          "summary": "Update Customer",
          "operationId": "update_customer_customers__customer_id__put",
          "parameters": [
            {
              "name": "customer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Customer Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Customers"
          ],
          "summary": "Delete Customer",
          "operationId": "delete_customer_customers__customer_id__delete",
          "parameters": [
            {
              "name": "customer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Customer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/campaigns/": {
        "get": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Get Campaigns",
          "operationId": "get_campaigns_campaigns__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/CampaignRead"
                    },
                    "type": "array",
                    "title": "Response Get Campaigns Campaigns  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Create Campaign",
          "operationId": "create_campaign_campaigns__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CampaignRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/campaigns/{campaign_id}": {
        "get": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Get Campaign",
          "operationId": "get_campaign_campaigns__campaign_id__get",
          "parameters": [
            {
              "name": "campaign_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Campaign Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CampaignRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Update Campaign",
          "operationId": "update_campaign_campaigns__campaign_id__put",
          "parameters": [
            {
              "name": "campaign_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Campaign Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CampaignRead"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Delete Campaign",
          "operationId": "delete_campaign_campaigns__campaign_id__delete",
          "parameters": [
            {
              "name": "campaign_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Campaign Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "CampaignCreate": {
          "properties": {
            "campaign_id": {
              "type": "integer",
              "title": "Campaign Id"
            },
            "campaign_name": {
              "type": "string",
              "title": "Campaign Name"
            },
            "start_date": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            },
            "end_date": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            },
            "budget": {
              "type": "number",
              "title": "Budget"
            },
            "status": {
              "type": "string",
              "title": "Status"
            }
          },
          "type": "object",
          "required": [
            "campaign_id",
            "campaign_name",
            "start_date",
            "end_date",
            "budget",
            "status"
          ],
          "title": "CampaignCreate"
        },
        "CampaignRead": {
          "properties": {
            "campaign_id": {
              "type": "integer",
              "title": "Campaign Id"
            },
            "campaign_name": {
              "type": "string",
              "title": "Campaign Name"
            },
            "start_date": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            },
            "end_date": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            },
            "budget": {
              "type": "number",
              "title": "Budget"
            },
            "status": {
              "type": "string",
              "title": "Status"
            }
          },
          "type": "object",
          "required": [
            "campaign_id",
            "campaign_name",
            "start_date",
            "end_date",
            "budget",
            "status"
          ],
          "title": "CampaignRead"
        },
        "CustomerCreate": {
          "properties": {
            "customer_id": {
              "type": "integer",
              "title": "Customer Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "age": {
              "type": "integer",
              "title": "Age"
            },
            "income": {
              "type": "number",
              "title": "Income"
            },
            "credit_score": {
              "type": "integer",
              "title": "Credit Score"
            }
          },
          "type": "object",
          "required": [
            "customer_id",
            "name",
            "age",
            "income",
            "credit_score"
          ],
          "title": "CustomerCreate"
        },
        "CustomerRead": {
          "properties": {
            "customer_id": {
              "type": "integer",
              "title": "Customer Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "age": {
              "type": "integer",
              "title": "Age"
            },
            "income": {
              "type": "number",
              "title": "Income"
            },
            "credit_score": {
              "type": "integer",
              "title": "Credit Score"
            }
          },
          "type": "object",
          "required": [
            "customer_id",
            "name",
            "age",
            "income",
            "credit_score"
          ],
          "title": "CustomerRead"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }